Project & Documentation/Project FaceApp/requirements.txt

# Minimal requirements (fallback uses OpenCV Haar cascades)
Flask>=2.0
opencv-python
numpy
Pillow
werkzeug
